Move to internal implementation AndroidPlayerSurfaceView
